+++ src/Press/Http/Filters/StatusFilter.php
@@ -19,11 +19,6 @@ class StatusFilter extends Filter
     ];
 
     /**
-     * @var bool
-     */
-    public $dashboard = true;
-
-    /**
      * @param Builder $builder
      *
      * @return Builder